predict/requirements.txt
tensorflow==2.17.0  
Flask
gunicorn
google-cloud-firestore
